Refactor trino2rel to extract parsetree package and remove unused class
