+++ coral-trino/src/main/java/com/linkedin/coral/trino/trino2rel/parsetree/UnhandledASTNodeException.java
@@ -3,7 +3,7 @@
  * Licensed under the BSD-2 Clause license.
  * See LICENSE in the project root for license information.
  */
-package com.linkedin.coral.trino.trino2rel;
+package com.linkedin.coral.trino.trino2rel.parsetree;
 
 import coral.shading.io.trino.sql.tree.Node;
